// src/webview/types/I18nDef.ts
export enum SupportedLocales {
  EN_US = 'en_US', // English (US)
  ES_MX = 'es_MX', // Spanish (Mexico)
  DE_DE = 'de_DE', // German (Germany)
  FR_FR = 'fr_FR', // French (France)
  IT_IT = 'it_IT', // Italian (Italy)
  JA_JP = 'ja_JP', // Japanese (Japan)
  KO_KR = 'ko_KR', // Korean (Korea)
  PL_PL = 'pl_PL', // Polish (Poland)
  PT_BR = 'pt_BR', // Portuguese (Brazil)
  RU_RU = 'ru_RU', // Russian (Russia)
  TR_TR = 'tr_TR', // Turkish (Turkey)
  ZH_CN = 'zh_CN', // Chinese (Simplified)
  CS_CZ = 'cs_CZ', // Czech (Czech Republic)
  UK_UA = 'uk_UA'  // Ukrainian (Ukraine)
}